requirements.txt
# Google Cloud Text-to-Speech Application Dependencies
# Following VIPER architecture pattern with Google Cloud Platform services

# Core Flask framework for API and web interface
Flask==3.0.0
Flask-CORS==4.0.0

# Google Cloud Platform SDKs
google-cloud-texttospeech==2.21.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-cloud-secret-manager==2.17.0

# Google Auth and API Core
google-auth==2.25.2
google-api-core==2.15.0
googleapis-common-protos==1.62.0

# Data validation and serialization
pydantic==2.5.2
marshmallow==3.20.2

# Configuration management
PyYAML==6.0.1
python-dotenv==1.0.0

# HTTP and API utilities
requests==2.31.0
urllib3==2.1.0

# Date and time handling
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Code quality and linting
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.8.0

# Security
cryptography==41.0.8
PyJWT==2.8.0

# Performance and caching
redis==5.0.1
gunicorn==21.2.0

# API documentation
flasgger==0.9.7.1
swagger-ui-bundle==0.0.9

# Utilities
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
itsdangerous==2.1.2

# Environment-specific dependencies
# Production
gevent==23.9.1
psutil==5.9.6

# Development
watchdog==3.0.0
ipython==8.18.1

# Type hints and annotations
typing-extensions==4.9.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# CLI framework and utilities
click==8.1.7
rich==13.7.0
colorama==0.4.6

# Data processing
pandas==2.1.4
numpy==1.26.2

# Audio processing utilities
pydub==0.25.1
mutagen==1.47.0

# File handling
pathlib2==2.3.7

# Network and HTTP
httpx==0.25.2

# Serialization
orjson==3.9.10

# Memory optimization
pympler==0.9

# Environment detection
distro==1.8.0

# Progress bars and interactive CLI
tqdm==4.66.1

# Configuration management
configparser==6.0.0

# Async support for CLI
asyncclick==8.1.3.4